fix: widen thread index to 64 bits in all kernels
All five one-thread-per-element kernels computed their global index as
blockIdx.x * blockDim.x + threadIdx.x. Both operands are unsigned int, so the
product is evaluated in 32 bits and wraps past 2^32 threads no matter that the
destination is uint64_t. Affected even numbers were then skipped by the GPU
and fell through to the single-threaded Phase 2 path, which presents as a hang
rather than a crash: --seg-size=4e9 finished in 0.9s, 4.5e9 ran past 300s.

Casting one operand yields a widening 32x32->64 multiply, a single instruction
on NVIDIA hardware, executed once per thread rather than in any inner loop.

Sites: goldbach.cu (count_unverified), sieve_kernel.cuh (large_prime_sieve),
phase1_kernel.cuh (scalar and transposed Phase 1), and single_check.cu, which
had the same pattern and was not in the original list. test_primality.cu
indexes with int against an int count and is bounded by construction, so it
is left as is.

With the kernels fixed, the 2^32 --seg-size guard added in the previous commit
is replaced by the real limits: per-device maxGridSize[0] for the widest
launch (one thread per even number, so seg_even_count/256 blocks), and the
existing free-memory check. On an RTX 5090 that is 549,755,813,632 by grid and
~1.24e11 by memory, so which one binds depends on the value.

Cost, measured with before/after runs interleaved, five each, 0 fallbacks
throughout:
  1e11  reported 0.8092 -> 0.8066 s (-0.32%),  wall 1.300 -> 1.284 s
  1e13  reported 89.898 -> 89.830 s (-0.07%),  wall 90.378 -> 90.312 s
Both deltas are negative, i.e. within noise: no measurable cost.

SASS confirms why. tiled_sieve_segment_kernel, 71-80% of runtime, indexes with
blockIdx.x alone and is byte-identical. large_prime_sieve_kernel swaps IMAD for
IMAD.WIDE.U32 with no change in instruction count. Only the transposed Phase 1
kernel grows, 80 -> 88 instructions, all in the prologue, because group_start *
64 becomes a 64-bit shift pair.

goldbach 4500000000 --seg-size=4500000000, the case that hung, now completes in
0.184s with 0 fallbacks.

+++ include/phase1_kernel.cuh
@@ -81,7 +81,7 @@ __global__ void goldbach_phase1_kernel(
     const uint64_t* __restrict__ p_batch, uint64_t p_batch_size,
     uint64_t* d_verified)
 {
-    uint64_t tid = blockIdx.x * blockDim.x + threadIdx.x;
+    uint64_t tid = (uint64_t)blockIdx.x * blockDim.x + threadIdx.x;
     if (tid >= seg_even_count) return;
     // Early return monotonic safety. This read is not atomic while other lanes
     // atomicOr the same word, but bits only ever go 0 -> 1, so a stale read
@@ -134,7 +134,7 @@ __global__ void goldbach_phase1_transposed_kernel(
     const uint64_t* __restrict__ p_batch, uint64_t p_batch_size,
     uint64_t* __restrict__ d_verified)
 {
-    uint64_t t           = blockIdx.x * blockDim.x + threadIdx.x;
+    uint64_t t           = (uint64_t)blockIdx.x * blockDim.x + threadIdx.x;
     uint64_t group_start = t * 64;                 // index of this thread's first even number
     if (group_start >= seg_even_count) return;
